Renaming Chapter 9 to Chapter 10 and vice-versa
