+++ CH09/.gitignore
@@ -1,8 +1,15 @@
 logs
+project/project
+project/target
 target
+tmp
+.history
+dist
 /.idea
+/*.iml
+/out
 /.idea_modules
 /.classpath
 /.project
-/.settings
 /RUNNING_PID
+/.settings
