+++ CH09/conf/routes
@@ -3,9 +3,8 @@
 # ~~~~
 
 # Home page
-GET     /                           controllers.Application.index
-GET     /text                       controllers.Application.text
+GET        /                    controllers.Application.index
+GET        /stream              controllers.Application.stream
 
 # Map static resources from the /public folder to the /assets URL path
-GET     /assets/*file               controllers.Assets.versioned(path="/public", file: Asset)
-GET     /webjars/*file              controllers.WebJarAssets.at(file)
+GET        /assets/*file        controllers.Assets.at(path="/public", file)
