+++ CH10/.gitignore
@@ -1,15 +1,8 @@
 logs
-project/project
-project/target
 target
-tmp
-.history
-dist
 /.idea
-/*.iml
-/out
 /.idea_modules
 /.classpath
 /.project
-/RUNNING_PID
 /.settings
+/RUNNING_PID
