+++ CH10/app/assets/javascripts/application.js
