+++ CH10/conf/routes
@@ -3,8 +3,9 @@
 # ~~~~
 
 # Home page
-GET        /                    controllers.Application.index
-GET        /stream              controllers.Application.stream
+GET     /                           controllers.Application.index
+GET     /text                       controllers.Application.text
 
 # Map static resources from the /public folder to the /assets URL path
-GET        /assets/*file        controllers.Assets.at(path="/public", file)
+GET     /assets/*file               controllers.Assets.versioned(path="/public", file: Asset)
+GET     /webjars/*file              controllers.WebJarAssets.at(file)
