+++ CH10/project/build.properties
@@ -1,4 +1,4 @@
 #Activator-generated Properties
-#Thu Jul 02 15:21:14 CEST 2015
-template.uuid=49a31253-85ed-4c4f-9041-f2f030591a8d
+#Thu Jul 23 09:07:38 CEST 2015
+template.uuid=a91771f5-1745-4f51-b877-badeea610f64
 sbt.version=0.13.9
